search-scripts_old/requirements.txt
werkzeug
tqdm
certifi==2019.6.16
chardet==3.0.4
elasticsearch==7.0.4
idna==2.8
joblib==0.14.0
mysql-connector-python==8.0.17
numpy==1.17.1
pandas==0.25.1
Pebble==4.4.0
protobuf==3.9.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
graypy==2.1.0
